Remove legacy incremental processing and simplify chunked refresh
Eliminates incremental and checkpoint-based processing logic, including related CLI options, models, and scripts. Updates chunked refresh to use a unified DataProcessor with timestamp-based staleness selection. Cleans up workflow, scripts, and config to remove unused incremental features and documentation.
