+++ scripts/format.sh
@@ -5,14 +5,15 @@ echo "Formatting Python code with black..."
 
 if [ -d ".venv" ]; then
     source .venv/bin/activate
-elif [ -d "venv" ]; then
-    source venv/bin/activate
+else
+    echo "Error: .venv directory not found. Run ./scripts/setup.sh first"
+    exit 1
 fi
 
-if ! command -v black &> /dev/null; then
+if ! python -m black --version &> /dev/null; then
     echo "Error: black is not installed. Run: pip install black"
     exit 1
 fi
 
-black swift_package_analyzer/ swift_analyzer.py
+python -m black swift_package_analyzer/ swift_analyzer.py
 echo "Code formatting complete."
